mv set convert and clean datum method to own files
